Split test-helpers so it works with frontend
