+++ packages/frontend/package.json
@@ -29,6 +29,7 @@
   },
   "devDependencies": {
     "@esp-group-one/eslint-config": "workspace:^",
+    "@esp-group-one/test-helpers-base": "workspace:^",
     "@testing-library/dom": "^9.3.4",
     "@testing-library/jest-dom": "^6.4.2",
     "@testing-library/react": "^14.2.1",
