+++ packages/test-helpers/src/index.ts
@@ -1,9 +1,6 @@
-import { helpers as types } from "@esp-group-one/types";
 import { helpers as db } from "@esp-group-one/db-client";
 
-const { IDS, OIDS, getLeague, getMatch, getUser } = types;
-
-export { IDS, OIDS, db, getLeague, getMatch, getUser, types };
+export { db };
 export * from "./db.js";
-export * from "./mock.js";
 export * from "./jest.js";
+export * from "@esp-group-one/test-helpers-base";
